🚚 move types into a common folder
