+++ test/e2e/scenario/rum/resources.scenario.ts
@@ -1,7 +1,7 @@
 import { createTest, html } from '../../lib/framework'
 import { sendXhr } from '../../lib/helpers/browser'
 import { flushEvents } from '../../lib/helpers/sdk'
-import { ServerRumResourceEvent } from '../../lib/serverTypes'
+import { ServerRumResourceEvent } from '../../lib/types/serverEvents'
 
 const REQUEST_DURATION = 200
